requirements.txt
python-hcl2 
ruamel.yaml
pyyaml
zss
pyvis
networkx